in preparation of a workshop release

Conflicts:
	applications/plugins/Compliant/odesolver/CompliantStaticSolver.cpp
	applications/plugins/Compliant/python/Compliant/sml.py
	applications/plugins/Flexible/python/Flexible/API.py
	framework/sofa/helper/system/FileSystem.cpp
	modules/SofaOpenglVisual/ColorMap.cpp

Former-commit-id: 4dcd9041e3d4a8aecc63b2cd00787979eaef899b

+++ modules/SofaBaseCollision/CubeModel.h
@@ -24,6 +24,7 @@
 ******************************************************************************/
 #ifndef SOFA_COMPONENT_COLLISION_CUBEMODEL_H
 #define SOFA_COMPONENT_COLLISION_CUBEMODEL_H
+#include "config.h"
 
 #include <sofa/core/CollisionModel.h>
 #include <SofaBaseMechanics/MechanicalObject.h>
@@ -116,7 +117,7 @@ public:
     void setLeafCube(int cubeIndex, std::pair<core::CollisionElementIterator,core::CollisionElementIterator> children, const sofa::defaulttype::Vector3& min, const sofa::defaulttype::Vector3& max);
 
 
-    unsigned int getNumberCells() { return elems.size();}
+    unsigned int getNumberCells() { return (unsigned int)elems.size();}
 
     void getBoundingTree ( sofa::helper::vector< std::pair< sofa::defaulttype::Vector3, sofa::defaulttype::Vector3> > &bounding )
     {
